feat: add Lite, non-customizable port of v2 theme
This is only meant to be used with extensions that don't have a support
for stylus-lang and/or external projects. As such, its purpose is to
provide you with the core functionality of Dark-WhatsApp without the
need to compile it on your own.

Compiling manually is something to work on for future updates. At this
moment, you can only convert the `wa.user.css` to `darkmode.css` for use
in Ferdi/Franz. More info will be in installation section of the wiki.

Partial support for #50.