chat: improve RTL msg tweaks and fix edge-cases
By using a slightly different approach, we're able to add padding to the
RTL messages and it will add the right amount of whitespace for text not
to collide with the timestamp.

This also removes tweaks in Chromium-based browsers, and by doing so,
the 'Read more' button is displayed inline with the text instead of
being pushed to the next row.

Huge thanks @E-RELevant for the help!