feat: add project logo to intro image
Related to #156.

Due to the goals of this project, we decided not to merge the PR and
improve upon the idea by using a logo that uses theme colors. It works
really well and looks far better than what we currently have.

Huge thanks to @jdruedaq for contributing and for the initial idea!