chore: fix the order of expanding helper variables
Strangely enough, the `{ md }, > div > div` will actually compile
successfully, yet produce no output, and it took me a _while_ to figure
that out.

A workaround is to put interpolation at the very end, and for whatever
reason it _just works_. However, it seems like it doesn't play well with
the `&:not({ ... })` selector so that one had to be split to two lines.