feat: implement custom kiss/pulse emoji options
This builds on top of the initial idea that Erel implemented, and gives
you the ability to customize which emojis have animations.

To disable individual emoji options, you can replace all the text with
the following characters:
- ''
- ()

To add new emojis, head over to Emojipedia [1] (or some similar
website), search for and copy the emoji(s) you want to use, then
surround it/them with single or double quotes. Using quotes is
mandatory, as well as space characters between chosen emojis.

Example:
- '🥰' '✨' '🔥'

P.S. All of this is all done via configuration menu [2], and not
directly in the source code.

1: https://emojipedia.org/
2: https://github.com/vednoc/dark-whatsapp/wiki/Configuration